Implementa funcionalidades e rotas para funcionamento da api

+++ src/api/Models/EntityModel/UserProjects/UserProject.cs
@@ -5,10 +5,10 @@ namespace api.Models.EntityModel.UserProjects
 {
     public class UserProject
     {
-        public long Id { get; set; }
-        public long ProjectId { get; set; }
+        public int Id { get; set; }
+        public int ProjectId { get; set; }
         public Project? Project { get; set; }
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public User? User { get; set; }
     }
 }
